[rdrawable] split attributes classes in different sources
